dom/indexedDB/test/test_setVersion.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function testSteps()
{
  const name = window.location.pathname;
  const description = "My Test Database";

  var request = indexedDB.open(name, description);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  var event = yield;

  var db = event.result;

  // Check default state.
  SimpleTest.is(db.version, "", "Correct default version for a new database.");

  const versions = [
    "",
    "1.0",
  ];

  for (let i = 0; i < versions.length; i++) {
    let version = versions[i];
    let request = db.setVersion(version);
    request.onerror = errorHandler;
    request.onsuccess = continueTest;
    event = yield;

    SimpleTest.is(db.version, version,
                  "Database version number updated correctly");
  }

  SimpleTest.executeSoon(finishTest);
  yield;
}

var gGen = testSteps();
function run_test()
{
  SimpleTest.waitForExplicitFinish();
  continueTest();
}

  </script>

</head>

<body onload="run_test();">
</body>
</html>
